full refactoring structure
