+++ ImageFeed/Controllers/ProfileViewController/ProfileViewPresenter.swift
@@ -9,19 +9,25 @@ import Foundation
 
 protocol ProfileViewPresenterProtocol {
     var view: ProfileViewControllerProtocol? {get set}
+
     func startObservingProfileImageChanges()
     func setupGradientAnimation()
     func didLogout()
 }
 
 final class ProfileViewPresenter: ProfileViewPresenterProtocol {
+
+    // MARK: - Public Properties
+
     weak var view: ProfileViewControllerProtocol?
     var profileImageServiceObserver: NSObjectProtocol?
-    var router: LogoutRouterProtocol = LogoutRouter()
+    var router: ProfileViewRouterProtocol = ProfileViewRouter()
+
+    // MARK: - Public Methods
 
     func setupGradientAnimation() {
         guard let view = view else { return }
-        ProfileHelper().setupGradientAnimation(view: view)
+        ProfileService().setupGradientAnimation(view: view)
     }
 
     func startObservingProfileImageChanges() {
