+++ ImageFeed/Controllers/ProfileViewController/ProfileViewRouter.swift
@@ -1,5 +1,5 @@
 //
-//  LogoutRouter.swift
+//  ProfileViewRouter.swift
 //  ImageFeed
 //
 //  Created by Artem Krasnov on 05.03.2024.
@@ -7,13 +7,19 @@
 
 import UIKit
 
-protocol LogoutRouterProtocol {
+protocol ProfileViewRouterProtocol {
     var view: ProfileViewControllerProtocol? { get set }
+
     func logout()
 }
 
-final class LogoutRouter: LogoutRouterProtocol {
+final class ProfileViewRouter: ProfileViewRouterProtocol {
+
+    // MARK: - Public Properties
+
     weak var view: ProfileViewControllerProtocol?
+
+    // MARK: - Public Methods
 
     func logout() {
         OAuth2TokenStorage.shared.deleteToken()
