+++ ImageFeed/Controllers/SplashViewController/TabBarController.swift
@@ -8,16 +8,22 @@
 import UIKit
 
 final class TabBarController: UITabBarController {
+
+    // MARK: - Overrides Methods
+
     override func awakeFromNib() {
         super.awakeFromNib()
+
         let storyboard = UIStoryboard(name: "Main", bundle: .main)
         let imageListViewController = storyboard.instantiateViewController(
             withIdentifier: "ImagesListViewController")
+
         let profileViewController = ProfileViewController()
         profileViewController.tabBarItem = UITabBarItem(
             title: nil,
             image: UIImage(named: "tab_profile_no_active"),
             selectedImage: nil)
+
         self.viewControllers = [imageListViewController, profileViewController]
     }
 }
