+++ ImageFeed/Services/ImageService/ImageLoadService.swift
@@ -12,14 +12,19 @@ protocol ImageLoaderProtocol {
     func loadImage(on url: String, completion: @escaping (Result<UIImage, Error>) -> Void)
 }
 
-final class ImageLoader: ImageLoaderProtocol {
+final class ImageLoadService: ImageLoaderProtocol {
+
+    // MARK: - Public Properties
+
+    let options: KingfisherOptionsInfo = [.cacheMemoryOnly, .transition(.fade(0.2))]
+
+    // MARK: - Public Methods
+
     func loadImage(on stringUrl: String, completion: @escaping (Result<UIImage, Error>) -> Void) {
        guard let imageURL = URL(string: stringUrl) else {
            completion(.failure(PhotoError.invalidURL))
            return
        }
-
-       let options: KingfisherOptionsInfo = [.cacheMemoryOnly, .transition(.fade(0.2))]
 
        KingfisherManager.shared.retrieveImage(with: imageURL, options: options) { result in
            switch result {
